+++ libmov/source/mov-stts.c
@@ -237,7 +237,7 @@ void mov_apply_ctts(struct mov_track_t* track)
     for (i = 0, n = 0; i < stbl->ctts_count; i++)
     {
         for (j = 0; j < stbl->ctts[i].sample_count; j++, n++)
-            track->samples[n].pts += (int32_t)stbl->ctts[i].sample_delta - dts_shift; // always as int, fixed mp4box delta version error
+            track->samples[n].pts += (int64_t)((int32_t)stbl->ctts[i].sample_delta - dts_shift); // always as int, fixed mp4box delta version error
     }
     assert(0 == stbl->ctts_count || n == track->sample_count);
 }
